requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
multidict==4.7.6
python-engineio==3.12.1
python-socketio==4.5.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
